Submitting a form with AJAX phpacademy
Submit a form with AJAX and fall back to normal form behaviour if
JavaScript is unavailable. We use the jQuery JavaScript library to
process the AJAX request. We also ensure that this can be re-used
without changing any JavaScript and the addition of more input fields
will just work. A great way to process form submissions with no page
refresh.